requirements.txt
aniso8601==9.0.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
Flask-PyMongo==2.3.0
gitdb==4.0.5
GitPython==3.1.13
httplib2==0.19.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
parse==1.19.0
parse-type==0.5.2
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
PyHamcrest==2.0.2
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
sentinels==1.0.0
six==1.15.0
smmap==3.0.5
Werkzeug==0.16.1
uWSGI==2.0.19.1
httpretty==1.0.5
behave==1.2.6
mock==4.0.3
mongomock==3.22.1
requests==2.25.1